bifangback/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-simple-history==2.12.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
h11==0.12.0
httpcore==0.12.2
httpx==0.16.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient @ file:///D:/mariadb/mysqlclient-1.4.6-cp38-cp38-win_amd64.whl
packaging==20.8
PyJWT==1.7.1
pyparsing==2.4.7
python-gitlab==2.5.0
pytz==2020.5
requests==2.25.1
rfc3986==1.4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
saltypie==0.14.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
termcolor==1.1.0
terminaltables==3.1.0
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.13.3